crypto: Move SerializedLegacyPublicKey to crypto
Now that the SerializedLegacyPublicKey has been re-arranged to remove
the strict dependence on PushBytes, it can be trivially moved to the
crypto crate.

Move SerializedLegacyPublicKey to crypto crate and re-export from
bitcoin.
Move associated test to crypto key module.
